Improve value output
Because pretty-print does a print for each value within a list, plus
for each formatting string, it can result in many calls to
repl-output-value with short strings -- which is inefficient and can
be visibly slow on the Emacs side.

To avoid this: For each current-print use for current-output-port a
pipe of unlimited size to accumulate all the pretty-printed bytes
and specials. Finally drain it using read-bytes-avail! to
consolidate runs of bytes (interrupted only by specials, if any)
up to a fixed buffer size.