Font-lock all output values and input, since last run
This allows a font-lock-fontify-region function, whether from classic
racket-mode or racket-hash-lang-mode, to font-lock output values (i.e.
from current-print) and interaction input -- provided they are related
to the most-recent run.

In some sense this aligns us with the DrRacket interactions pane which
is cleared before each run (a behavior option we also support if users
prefer). However we also allow retaining older interactions in the
buffer for users (like me) who prefer that. The old stuff is just "off
limits" and can't be font-locked richly if it hasn't already.

Now that we get structured output from the back end, and can
distinguish current-print values from miscellaneous stdout/stderr
output, it's fine to let classic racket-mode font-lock go to town on
them. So this commit livens up the appearance compared to previous
commits where we had to tone things down to err on the safe side, not
only for arbitrary hash-langs but also for things like issue