racket-repl: Add back async for back end startup scenario
With commit ac6a007 I overlooked the scenario where a racket-repl or
racket-run command is issued when the back end is stopped. We don't
want to block Emacs while it starts.

However we can keep some of the simplification, replacing the
repl-start helper function with an "ensure" style function that takes
a continuation callback.

- If the REPL buffer doesn't yet exist, it's created and put into
racket-repl-mode.

- If the buffer exists but has no live session (yet, or a previous
session was ended), then start one. If that means starting the back
end, show a message in the REPL buffer to explain the wait. The
continuation is called from the repl-start response callback.

- And of course if the buffer and session both are ready, then just
call the continuation.

One wrinkle is if any before-run hooks issue a command, they'll have
already begun the back end startup process and users won't see the
message. I changed one such example, racket-debug-disable, to only
issue its command when the back end is running (otherwise it's N/A).