Don't require Racket 8.4 just for current-get-interaction-evt
As another iteration on issue #671: Use current-get-interaction-evt
when available. Otherwise use yield and warn user that GUI programs
might not work properly before Racket 8.4.

I think this might be acceptable because:

- Some users of older Racket versions might in fact be using them on
systems where gui-lib isn't installed, or if installed can't
initialize (headless) -- scenarios we explicitly support -- in which
case this is moot.

- Some users will be OK with updating to a newer version of Racket.

- In my very limited testing GUI programs do work fine when we use
yield (but if not, see above).

In any case we keep the minimum hard-required version of Racket back
down at 6.12.