Assume only plain syntax initially
Avoids problems with e.g. paredit-mode being active before we get any
on-new-lang notification.

(That can happen when paredit-mode is on a racket-mode-hook. Because
racket-hash-lang-mode is derived from racket-mode, it runs those
hooks, too. For this plus other reasons, the more I think about rhlm
probably should /not/ be /derived/ from rm. It may inherit various
settings. Instead may it should derive directly from prog-mode, and
common functionality like keymap and menus could be handled by a
minor-mode ("mix-in") or just setting via a shared helper function.
But all that is for some future commit.)