Default racket-xp-binding-import-face to 'default face
This is closer to classic racket-mode.

Furthermore it avoids (at least by default) calling attention to
exactly where check-syntax thinks that end of the arrow is -- which
isn't always intuitive, e.g. as @samth pointed out for rhombus import
aliases.

Probably we can do a better job for such arrows with pdb, have
racket-pdb-mode supply a similar "semantic faces for bindings"
feature, and suggest using that instead with racket-hash-lang-mode.