Handle full set of module-language values; #673
I misread the docs; the value is not necessarily a symbol. It's a
module-path, or syntax whose datum is a module path, or even an
arbitrary procedure (which I assume is supposed to return one of
those).

This pushes me into supplying the module-language to the Emacs user as
a string (because Racket module paths in general are not necessarily
Racket symbols, much less Emacs symbols) but I think that's fine for
end user config purposes. A few early testers might have to change
this, is the only gotcha.