Don't use xvfb-run
I noticed this while diff-ing the hash-lang branch to master.

As best I can recall, using xvfb-run was necessary only prior to the
introduction of textoid<%>, so I'm removing that now and we'll see.

Also: Because the linux job runs the racket and emacs tests using
separate steps, also do so for the windows job. (Although it should be
the same result either way, be consistent.)