Fix: Ensure REPL configure from edit buffer happens on first run
This requires waiting to do it until the repl buffer is created and
sufficiently initialized -- but before issuing any command to start a
repl session.

While we're add it, change this from a hook to a direct function call.
There's nothing "hooky" about this, we just need to avoid a mutual
require, and we can do that the usual way with declare-function and
autoload.