Don't relay write-special for stdout/stderr; fixes #677
Although we handle write-special in a temporary values port in
print.rkt, for convertible image values, we don't need to generally
for current-output-port or current-error-port.

We were already ignoring non-image special values in the Emacs front
end. So handling write-special in the general ports in the back end
only means we might try, unnecessarily, to elisp-write some arbitrary
Racket value, and sometimes fail. (Whatever the student langs and
DrRacket are doing in this regard with write-special, beyond images,
we can ignore for now, and maybe forever.)