Ignore new <span class="button-group"> elements
These use "link" and "i" emoticons that...

(a) render as \nnn byte strings, because we're using
insert-file-contents-literally, and I don't have time now to figure
out why we use that instead of insert-file-contents, and how to use
the latter for encoding while avoiding whatever other behavior we
wanted to avoid.

(b) in a browser only show when the user hovers with the mouse, but in
our case would appear always, after every main section heading, which
is nosiy.

Someday/maybe we should figure out a way to present the hidden
information -- a "link here" URL, as well as the scribble source doc
URL on GitHub -- that's appropriate for our UX. And then deal with the
various ways this has been presented in Scribble HTML over time.