racket-repl-output: buffer/"batch" stdout/stderr
Performance can be bad at either of two extremes: Very many outputs,
or very large outputs. This can arise with stdout or stderr outputs:
e.g. (for ([_ huge-number]) (print something)) or (print huge-thing).
Use a manager thread to mediate and buffer/batch.

(Note that very long lines can still be slow in any Emacs buffer, but
(a) that can be mitigated by so-long and IIUC Emacs 29.1, and (b)
anyway that's not a problem we were contributing.)