racket-repl: Add output filter and use for ansi-color; fixes #688
Add racket-repl-output-filter-functions, which is similar to
comint-preoutput-filter-functions but restricted to stdout and stderr
kinds of output; it doesn't affect any other kinds like prompts,
input, etc.

Default this to use ansi-color-apply, which replaces SGR escapes with
colors. A user could remove this to see raw SGR escapes, or change it
to ansi-color-filer-apply to remove them.

Also: Add a racket-before-run-hook item to reset the state, so that an
escape from a previous run doesn't affect output for subsequent runs.

TL;DR: We get ansi coloring, but without defects in comint like the
coloring bleeding inappropriately into other kinds of output or
persisting inappropriately.