racket-repl: Re-implement as just running a file; closes #689
Narrowly, this fixes the kind of problem as in #689 where a setting
like racket-pretty-print only takes effect via a run.

Broadly, this makes this command follow the the main intended use case
-- a REPL is associated with running a file. Now the command is just a
shortcut for running a file with a certain name (and creating that
file if it doesn't exist).

Also, this fits better in a hash-lang world. If say someone mostly
works with #lang rhombus, and wants "give me a quick REPL" to use
that, now they can instead of getting #lang racket/base.

Finally, this lets the doc string explain that the REPL is always
just about running some file. If a user wanted to have "several init
files", well, they're really talking about running several source
files frequently, which of course they can already do. And if they
want a quick key binding for each one, that's an easy customization
nearly any end user can do for themselves.

TL;DR: Years ago when some folks asked for a racket-repl command, this
is probably how I should have done it: A trivial variation of
racket-run.