Add list-racket-packages and describe-racket-package
Issue #710 planted the idea.

Largely follow the UX used by Emacs for its list-packages (a tabulated
list view) and describe-package (a help buffer with details).

In latter, "link all the things": can visit paths, URLs, and details
of dependency packages.

Append some general package system configuration. On the one hand this
is weird, because the information is not details about a specific
package. On the other hand, we distinguish it in the UX, and it seems
easier (for users) to see it there than needing to discover and issue
some additional command.

Support package names as a distinct 'racket-package completion
category, both to avoid things like marginalia mistakenly classifying
and annotating our choices, as well as to allow users to customize
completion styles.

Do package operations in each back end, to support a mix of Racket
version installations, both remote and local.