Various changes
- Add a button to query online for rendered documentation (more
directly than visiting the catalog page to get a link there). This
information isn't provided by the local package API, which is why we
do it on-demand with a button rather than proactively do an HTTPS
request for every package details buffer.

- Replace blank string values with #f, which serialize to nil on the
Emacs front end, so we can omit showing detail labels with no values.
One example where this helps is with planet packages, which usually
seem to have blank Catalog and Author strings.

- Improve more doc strings.