Add racket-package-refresh command; remove auto refresh
The previous commit would refresh the local catalog on every startup
of the Racket Mode back end.

On reflection this seems wrong. Some users might not want a connection
to made.

Instead supply a command for users to do this explicitly, when they
choose. Downside is the results will be poor until they do this at
least once, but we document this. Anyway this is analogous to
`package-refresh-contents` for Emacs packages, and the idea here was
already to largely follow the same UX.