Better support context-menu-mode
The TL;DR is not to assume `point` is the relevant position. When
`this-command` came from a mouse event, use the mouse cursor
position (else use point as usual).

Although some things like xref define plain and mouse flavors of all
the commands, that seems like a lot of busywork. Instead follow the
example of the recent the magit 4.0 update: Define wrappers for things
like `point` and `thing-at-point` to use instead.