Improve input method customization
Add two customization variables from which the quail input method is
defined: racket-input-prefix and racket-input-translations.

The prefix defaults to \ and avoids common conflicts.

Add a minor mode to activate the input method: racket-input-mode. The
old racket-unicode-input-method-enable is now just a defalias for
that.

Add a plain command, as an alternative to the input method as a way to
use racket-input-translations: racket-insert-symbol.

Closes #718.
Closes #723.

Related to already-closed issue #717: This commit reverts the change
from "oint" to "Oint", on the theory that it's unnecessary now with
the prefix. Even if someone wanted to have no prefix (like before),
they could easily M-x customize that, and they could easily M-x
customize "oint" back to "Oint" or whatever they want. TL;DR the whole
point of this commit is to enable most such changes to be a
customization instead of a PR.