Further improve use with eldoc-box
1. Enable truncate-lines, as does racket-describe-mode. Avoids extra
line breaks and line continuation marks in the fringe. Example where
this helps: Set racket-xp-eldoc-level to 'complete, and get a hover
for the Racket require or provide forms.

2. Instead of copy-pasta-ing most of eldoc-box-buffer-setup, actually
call it, but within a cl-letf that "no-op"s visual-line-mode and
eldoc-box-buffer-hook. Although this may seem too clever, it should be
more resilient to eldoc-box-buffer-setup changing internally in
beneficial or neutral ways; no ongoing copy-pasta. (In the less-likely
worst case where it adds some new undesired behavior, sure, we will
have to neuter that using a technique like cl-letf.)