Add some basic smoke tests for doc-index-lookup
These are a few basic tests intended to check that things still work
with the "classic" index structs, as well as with the new ones.

Also, when rhombus is installed, we can add a few more tests, to see
e.g. if we get more interesting kinds than "value" for it, such as
"regular expression operator".

Note that these rhombus tests won't run on CI unless we change CI to
install the rhombus package, which I'm considering but would be a
future commit.