Simplify; abandon the agda-input ideas
After documenting and dog-fooding, I've realized the agda-input design
is great -- for something other than what we're doing.

It would be great for an input method that is designed to inherit from
one or more other, "basic" input methods. For that, a tweak-all
function that composes combinators is necessary and worthwhile. But
for our basic input method, it's overkill, and just makes it harder
for us to document and for users to customize (AFAICT).

Effectively this brings things back to an earlier commit, where I was
experimenting doing this via abbrev-mode instead of an input method.
Although abbrevs won't work, because they're effectively limited to
alphanumeric abbreviations, and this commit is using an input method,
the basic flavor is the same: One customization var for a prefix, and
another for a list of translations. And the translations can be used
both in an "automatic" mode (via the input method) as well as a
command, which has a nice completing-read UI with annotations.