Use a make- function instead of apply-partially
Also the PROP arg can now be optional.

Not only is this simpler for callers, we can get better control over
the lifetime of the max-widths variable. In other words, when people
are scrolling, if the columns get wider, and they scroll back up, they
remain at that new width, instead of shrinking again -- which is a
tiny detail but I think it's slightly nicer.