generate.el: Improve reference.org creation
1. Use "insert style" -- rather than build up big strings to insert,
insert pieces into a buffer as we go.

Although that's often best practice in Emacs, for efficiency (often
the best string API in Emacs is the buffer API) that's mostly N/A for
this doc build process. Here I just think the code clearer and simpler
that way. Although yeah it's imperative not functional, output is
imperative.

2. Simplify (re)creation of reference.org file. Instead of find-file,
just use a temp buffer and write-region.

3. Simplify the key binding grouping added in commit db26892. Rather
than group using an intermediate hash-table, group when building the
association list.

4. Shorten the namespace prefixes from "racket-generate--" to
"racket--". (I'm not sure we need prefixes, at all, given this is only
used from Emacs --batch.)