Improve racket-repl after-change behavior; fixes #731
Reduce the number and extent of calls to after-change-functions.

- Use inhibit-modification-hooks when using remove-text-properties on
rear-nonsticky text properties for the entire buffer. Our back end
hash-lang object doesn't need this huge, meaningless change. Nor do
things like jit-lock-after change. This mitigates the proximate cause
of #731.

- Consolidate newline+insert into insert (the "fresh line" stuff).

- Non-substantive: Rewrite cl-case as pcase.

Also, in racket-hash-lang.el, some non-substantive changes:

- Fix prose in doc string for racket--hash-lang-repl-buffer-string.

- Rename racket--hash-lang-after-change-hook to
racket--hash-lang-after-change -- for brevity, but also because
technically it's not a value for a hook, it's a value for the variable
`after-change-functions`.