Improve tidy-requires
Motivated by https://github.com/jackfirth/resyntax/issues/432, as well
as inspired by racket-review's used of syntax classes.

Define syntax classes closely following the require grammar. Use them
to drive the matching and simplification, producing attribute values
that are `sortable` structs, from which we reconstruct the syntax.

Because the macro-debugger-lib drop analysis works in terms of <module
phase> tuples, the root-module-path syntax class is the locus -- it
may produce #f instead of a `sortable` to indicate elision. This can
percolate up through other specs, since they all build directly or
transitively on root-module-path.

Simplify gratuitous specs either reducing to the simpler equivalent
spec or eliding entirely. e.g. (only-in m) => m or (combine-in) => #f.

No longer "explode" multi-in forms when de-normalizing and attempt to
recover them when normalizing. Instead: Preserve multi-in forms,
modifying only for drops. In simple cases like (multi-in a b (c d)),
we can modify the multi-in form itself. In the general case -- any
Cartesian product -- we must wrap in a subtract-in spec.

Improve some doc strings in substance, and prefer active to passive
voice.