Fix use of paredit-space-for-delimiter-predicates; fixes #736
All along we had been using `add-hook` to update paredit's
paredit-space-for-delimiter-predicates.

Commit 3a9675d started supply `t` for the optional LOCAL argument to
add-hook, which is the correct way set a local value for hook
variables.

The problem is, paredit-space-for-delimiter-predicates was never a
hook variable -- paredit doesn't use it via some run-hook variant, and
so it chokes on the `t` value.

Instead just setq-local it in each of our major modes, which should be
fine because they are major modes.