racket-describe completing-read initial/default
The completing-read docs say that INITIAL-VALUE is deprecated in favor
of supplying DEFAULT. Justification: The user can just M-n to load the
value at the prompt from "future history".

Although I had started to follow that advice, after dog-fooding this
for awhile as a user I've found it annoying. Furthermore, M-n
sometimes isn't sufficient to make completion candidates refresh. At
least in my experience, with vertico, you also need to C-e to move
point to the end. That is /super/ annoying, and probably super
confusing to users who aren't me.

This commit reverts back to supplying the Racket identifier at point
as the INITIAL-INPUT arg to completing-read.