Set current buffer in racket-describe-bookmark-jump
bookmark-handle-bookmark expects our handler to change the current
buffer, as does bookmark-default-handler.

Doing so ensures it shows the bookmark in a racket-describe-mode
buffer.

Otherwise it will jump to the bookmark position in whatever buffer
happens to be current. (I didn't notice the problem originally because
I only tested using the bookmark-jump command only in a
racket-describe-mode buffer.)

Also: Delete an unused declare-function.