racket-send-last-sexp: Don't insert \n; fixes #745
Note: This drops an option that "echoes" the expression to the REPL.
Doing so proactively is wrong for the same reason it's wrong to insert
a newline proactively. To revive such a feature, the back end would
need to hold onto this expression text in the interactions queue, and
send it back later using repl-output-comment just prior to the
repl-output-value. At the moment this seems like too much work for a
an option that I don't know if anyone is even using, so I'm taking the
approach of waiting to see if anyone actually notices or cares.