hash-lang: Use electric-pair-local-mode; fixes #747
Ditch our custom code to handle a lang supplying multi-char
delimiters.

Use electric-pair-local-mode by default, configuring it with a lang's
paren-matches and quote-matches, provided they are single characters
-- ignore multi char delimiters.

This is much simpler. Although it doesn't handle Rhombus '( and )'
delimiters directly, for example, the user can type '( to get '() and
then type )' and it works as expected.