Handle command line args to racket-program/backend
The documentation already had examples of configurations like

    (racket-add-back-end "/ssh:headless:~/gui-project/"
                         :racket-program "xvfb-run racket")

This would (accidentally?) work for remote backends, because the way the SSH
command is invoked ends up applying shell word splitting on the program.

However, when I wanted to use this locally something like this:

    (racket-add-backend "/project/"
                        :racket-program "podman
                                         run --rm --interactive
                                         --volume=/project/:/project/:z
                                         --workdir=/project/
                                         […]
                                         docker.io/racket/racket/8.9-full"
                                         "racket")

This wouldn't work because the `executable-find' call would try to find the path
to an executable with that full string as a name.

This commit adds the option of defining `racket-program` (or the back-end
`:racket-program` option) as a list rather than a string, in which case the
first element will be used as the command to call, and the rest of the elements
as command line arguments.