Don't font-lock logger output
Commit a23104e set font-lock-keywords-only. On reflection, it doesn't
make much sense to use font-lock at all in the logger output buffer.
Instead the back end could supply structured data (as opposed to a
string), and was can apply face properties directly when inserting.
This should be faster.

Similarly, insert a text property to let us find the start of each
item, and update the previous/next item commands to use that instead
of a regexp search.

Note: This uses the compat package to supply text-property-search
functions on older versions of Emacs.