Fix problems with racket-project-root; closes #752
1. Use project-root -- or, in older Emacs (car project-roots) -- to
extract the directory from the value returned by project-current.

Don't use cdr, which assumes a representation not necessarily true for
all project back ends.

2. Stop using vc-root-dir.

AFAICT, this was never appropriate, because it works in terms of the
current buffer -- using it ignores the FILE argument supplied to
racket-project-root. WTH was I thinking, years ago??

This simplifies racket-project-root to using the popular third-party
package "projectile" when available, else the built-in "project"
library when it can return a directory, else a final fallback of using
file-name-directory.