Interim commit working on issues related to lang rhombus
Separate the handling of (a) debug point expressions from (b) wrapping
REPL input in a let-syntax. Although both share the notion of making
locals visible within a user input REPL expression, they differ.

Debug point expressions will be explicitly a Racket language
expression, regardless of the user program lang.

REPL input, of course, must be lang-specific, so we must limit any
let-syntax (to enable make-set!-transformers) or let-values (for
read-only locals) to what the lang understands. And rhombus for
example has a very restricted #%top-interaction which basically means
we can't wrap anything, ever.

TODO: As of this commit there is still a problem with e.g. rhombus due
to `multi` being an unknown identifier -- a namespace or lexical
context problem??