Fix issues with debug REPL and rhombus
1. Proactively namespace-require the file module's language. This
replaces our hardcoded namespace-require of racket/base. This avoids
unknown identifier errors for `multi`.

2. Avoid a let-syntax wrapper of input, when not the default
read-interaction handler. Note that this doesn't really "fix" things
in the sense of making locals available for rhombus, not even
read-only, but it avoids errors from rhombus #%top-interaction when
the outer form isn't multi` or `block`.