Drop debugger REPL; add command to set local variable
The debugger REPL is too fraught with complications re how to wrap any
arbitrary lang expression in a let-syntax -- especially rhombus with
its restricted `#%top-interaction`.

Drop the debugger REPL.

Do add a command to set local variables, which is one of the few
practical things the debugger REPL could be used for.