racket-hash-lang-mode: Improve/enhance range indent; closes #759
Instead of binding <tab> to the usual indent-for-tab-command, use a
new command which persists any active region, making it less painful
to cycle among indents. (Persisting an active region requires us to
control the entire dynamic extent of the command and is impossible via
the indent-region-function approach. But do still supply such a
function for possible other situations where we want plain old
indent-region to keep working.)

Support the new drracket:range-indentation/reverse-choices, adding a
command bound to <backtab> for this.

On the back end, rename the hash-lang sub-command indent-region-amounts
to indent-range-amounts and give it a reverse? argument. Simplest way
to expose the non-reverse and reverse flavors.