Improve back end "type" command for r6rs; fixes #760
1. When a resolved module path doesn't exist, and it has a ".rkt"
extension, also try with ".ss". The standard module name resolver may
have changed it to ".rkt", so we must try to recover the original
".ss". This prevents the reported error message.

2. To the "find signature" code, add a special case for r6rs `library`
forms. This allows the example to work. (However, keep in mind that
all of this code to find signatures or definitions is very much ad hoc
and language dependent. It is the best that can be used in the REPL
environment. By contrast, in an edit buffer racket-xp-mode uses the
more principled and language agnostic approach of
drracket/check-syntax, with more useful and reliable results.)