Let next-error handle test coverage failures
While working on #761, I realized the test coverage UX could be better
-- specifically, why can't we navigate among missing coverage blocks
using the next-error UX.

As part of that, consolidate reporting of uncovered expressions so
that we don't report expressions that are mere subsets (this was
hidden by overlapping overlays but is obvious when reporting each
textually).

Do still also create the overlays in the buffer -- although IMHO
that's now redundant or even annoying, some people might be
familiar/comfortable with that status quo presentation.